package posposet_test renamed to posposet for private fields testing
