+++ src/posposet/node_hash_test.go
@@ -1,10 +1,8 @@
-package posposet_test
+package posposet
 
 import (
 	"math/rand"
 	"testing"
-
-	"github.com/Fantom-foundation/go-lachesis/src/posposet"
 )
 
 func TestAddress(t *testing.T) {
@@ -14,7 +12,7 @@ func TestAddress(t *testing.T) {
  * Utils:
  */
 
-func FakeAddress() (a posposet.Address) {
+func FakeAddress() (a Address) {
 	_, err := rand.Read(a[:])
 	if err != nil {
 		panic(err)
